fix(shaders): remove fp16 overflow hazards in sky shaders
Re-audited every hardcoded `precision mediump float;` fragment shader after
c6f9335 (moon disc invisible on mobile, fp16 overflow in the Bayer dither) for
the same class of bug. No further overflow hazard found - moon.ts's mod-4
reduction already closes it, and groundHaze/starField only touch bounded
world-Y and varying values.

moon.ts, moonHalo.ts, shootingStar.ts switch to highp anyway: each covers a
handful of screen pixels, so mediump's mobile perf win is unmeasurable, and
highp removes the residual risk of a future edit reintroducing an unreduced
large term. groundHaze.ts and starField.ts keep mediump - they cover the whole
sky every frame, so the perf win is real and the scan is clean. cityInstanced
untouched (no qualifier, renderer default, out of scope).

Byte-identical on desktop by construction (ANGLE/desktop GL already promote
mediump to fp32). Needs a phone check on the three highp shaders.

+++ lib/shaders/moonHalo.ts
@@ -12,7 +12,12 @@ export const moonHaloVertexShader = /* glsl */ `
 `;
 
 export const moonHaloFragmentShader = /* glsl */ `
-  precision mediump float;
+  // highp (2026-09-05 precision audit): no hazard found (all math is bounded UV-space
+  // distances/dots, div-by-zero already guarded), but this billboard covers a handful
+  // of screen pixels — mediump's mobile perf win is unmeasurable, so there is no
+  // upside to the residual risk. highp is the desktop default anyway. Needs a phone
+  // check. See wiki/notes/decision-shader-precision-policy.md.
+  precision highp float;
 
   uniform vec3 uColor;
   uniform float uInnerRadius; // 0..0.5 — radius where falloff begins (full opacity inside)
